FIX various fixes - size_t vs. int mismatches - few incorrect == vs = - few missing return values - few incorrect logging calls

+++ src/asset/ModelMesh.h
@@ -58,7 +58,7 @@ public:
     bool m_valid = false;
 
 protected:
-    int m_triCount = 0;
+    size_t m_triCount = 0;
     std::vector<glm::uvec3> m_tris;
     std::vector<Vertex> m_vertices;
     std::vector<Material> m_materials;
